+++ src/bittytax/conv/parsers/binance.py
@@ -476,7 +476,7 @@ def _parse_binance_statements_row(
         "Swap Farming Rewards",
     ):
         data_row.t_record = TransactionOutRecord(
-            TrType.STAKING,
+            TrType.STAKING_REWARD,
             data_row.timestamp,
             buy_quantity=Decimal(row_dict["Change"]),
             buy_asset=row_dict["Coin"],