+++ src/bittytax/conv/parsers/qtwallet.py
@@ -136,7 +136,7 @@ def _parse_qt_wallet_row(
         )
     elif row_dict["Type"] == "Masternode Reward":
         data_row.t_record = TransactionOutRecord(
-            TrType.STAKING,
+            TrType.STAKING_REWARD,
             data_row.timestamp,
             buy_quantity=amount,
             buy_asset=symbol,
@@ -209,7 +209,7 @@ def parse_vericoin_qt_wallet(
         )
     elif row_dict["Type"] == "Stake":
         data_row.t_record = TransactionOutRecord(
-            TrType.STAKING,
+            TrType.STAKING_REWARD,
             data_row.timestamp,
             buy_quantity=Decimal(row_dict["Amount"]),
             buy_asset=symbol,