+++ CHANGELOG.md
@@ -260,6 +260,7 @@
 - Deribit parser: use Margin-Fee-Rebate instead of Fee-Rebate for futures.
 - KuCoin parser: use Margin-Fee-Rebate instead of Fee-Rebate for futures.
 - Accounting tool: increased precision for Decimals from 28 to 30 digits.
+- Conversion tool: parsers use Staking-Reward instead of Staking (deprecated).
 ### Removed
 - Conversion tool: removed merge parser for Coinbase/Coinbase Pro.
 - Conversion tool: removed filename "is a directory" message.